+++ src/App.vue
@@ -1,7 +1,8 @@
 <script setup lang="ts">
-import { inject, ref } from "vue";
+import { inject, ref, onMounted } from "vue";
 
 import Breadcrumb from "./components/Breadcrumb.vue";
+import Form from "./components/Form.vue";
 
 import { Client } from "./client/Client";
 import { ISiyuan } from "./types/siyuan/siyuan";
@@ -11,14 +12,36 @@ const client = inject("client") as Client;
 const siyuan = inject("siyuan") as ISiyuan;
 const data = inject("data") as IData;
 const editable = ref(false);
+
+/* 调整高度 */
+function changeHeight(): void {
+    const height = document.getElementById("main")?.scrollHeight;
+    if (height && data.element) {
+        data.element.style.height = `${height + 2}px`;
+    }
+}
+
+/* 重新加载 */
+function reload(): void {
+    globalThis.location.reload();
+}
+
+/* 组件挂载 */
+onMounted(() => {
+    console.log("APP.onMounted");
+    setTimeout(changeHeight, 0);
+});
 </script>
 
 <template>
-    <a-layout class="layout">
+    <a-layout
+        id="main"
+        class="layout"
+    >
         <a-layout-header class="header">
             <a-row>
                 <a-col flex="none">
-                    <breadcrumb
+                    <Breadcrumb
                         :paths="data.paths"
                         :hpaths="data.hpaths"
                     />
@@ -27,25 +50,44 @@ const editable = ref(false);
                     <div>auto</div>
                 </a-col>
                 <a-col flex="none">
+                    <a-button
+                        @click="reload"
+                        :title="$t('reload')"
+                        size="mini"
+                        type="outline"
+                    >
+                        <template #icon>
+                            <icon-refresh />
+                        </template>
+                    </a-button>
+                    <a-divider direction="vertical" />
                     <a-switch
                         v-model:model-value="editable"
                         :title="$t('editable')"
+                        style="margin-top: -4px;"
                         type="circle"
                     >
                         <template #checked-icon>
-                            <icon-edit />
+                            <icon-check />
                         </template>
                         <template #unchecked-icon>
-                            <icon-eye />
+                            <icon-close />
                         </template>
                     </a-switch>
                 </a-col>
             </a-row>
         </a-layout-header>
         <a-layout>
-            <a-layout-content>Content</a-layout-content>
+            <a-layout-content>
+                <Form
+                    @updated="changeHeight"
+                    :client="client"
+                    :data="data"
+                    :editable="editable"
+                />
+            </a-layout-content>
         </a-layout>
-        <a-layout-footer>Footer</a-layout-footer>
+        <!-- <a-layout-footer>Footer</a-layout-footer> -->
     </a-layout>
 </template>
 
@@ -62,7 +104,7 @@ const editable = ref(false);
     }
     :deep(.arco-layout-header),
     :deep(.arco-layout-footer) {
-        padding: 0.5em;
+        padding: 0.25em 0.5em;
     }
 
     :deep(.arco-layout-sider) {