+++ src/components/Breadcrumb.vue
@@ -1,20 +1,24 @@
 <script setup lang="ts">
+import { computed, UnwrapNestedRefs } from "vue";
 import { BreadcrumbRoute } from "@arco-design/web-vue";
 
 const props = defineProps<{
-    paths: string[]; // 文档路径
-    hpaths: string[]; // 可读路径
+    paths: UnwrapNestedRefs<string[]>; // 文档路径
+    hpaths: UnwrapNestedRefs<string[]>; // 可读路径
 }>();
 
 /* 面包屑路径 */
-const end = Math.min(props.paths.length, props.hpaths.length);
-const routes: BreadcrumbRoute[] = [];
-for (let i = 0; i < end; ++i) {
-    routes.push({
-        label: props.hpaths[i],
-        path: props.paths[i],
-    });
-}
+const routes = computed(() => {
+    const routes: BreadcrumbRoute[] = [];
+    const end = Math.min(props.paths.length, props.hpaths.length);
+    for (let i = 0; i < end; ++i) {
+        routes.push({
+            label: props.hpaths[i],
+            path: props.paths[i],
+        });
+    }
+    return routes;
+});
 
 /* URL 生成 */
 function customURL(paths: string[]): string {