+++ src/types/siyuan/renameDoc.d.ts
@@ -1,0 +1,13 @@
+// /api/filetree/renameDoc
+
+export interface IPayload {
+	notebook: string;
+	path: string;
+	title: string;
+}
+
+export interface IResponse {
+	code: number;
+	msg: string;
+	data?: any;
+}