+++ src/utils/language.ts
@@ -16,3 +16,26 @@ export function mapLang(lang: string): string {
     }
     return lang;
 }
+
+/* 映射地区 */
+export function mapLocal(local: string): string {
+    switch (local) {
+        case "zh_CN":
+        case "zh_CHT":
+            local = "zh";
+            break;
+        case "es_ES":
+            local = "es";
+            break;
+        case "fr_FR":
+            local = "fr";
+            break;
+        case "en_US":
+            local = "en";
+            break;
+        default:
+            local = local.split('_').shift()!;
+            break;
+    }
+    return local;
+}